moved welcome modal to own file, removed dead files
