--- src/components/AdminDashboard/FormResponseIncidents.js
@@ -1,10 +1,0 @@
-import React, { useState, useEffect } from 'react';
-
-const FormResponseIncidents = () => {
-  const [formResponses, setFormResponses] = useState([]);
-
-  useEffect(() => {}, []);
-  return <div></div>;
-};
-
-export default FormResponseIncidents;
