--- src/components/AdminDashboard/Incidents.css
@@ -1,160 +1,0 @@
-.dashboard-top-flex {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: start;
-  -ms-flex-pack: start;
-  justify-content: flex-start;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-}
-
-.dashboard-top-flex div {
-  margin: auto 1%;
-}
-
-.dashboard-top-input {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: start;
-  -ms-flex-pack: start;
-  justify-content: start;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  width: 7%;
-}
-
-.dashboard-top-input input {
-  margin-left: 2%;
-}
-
-#select-all-label {
-  margin-left: 5%;
-}
-
-.dashboard-top-approve-reject {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  width: 50%;
-}
-
-.dashboard-top-approve-reject button {
-  margin: auto 2%;
-}
-
-.approve-reject-select {
-  border: 1px solid #e3e8ee;
-  background-color: white;
-  border-width: 2px;
-  border-style: outset;
-  border-radius: 5px;
-  margin-right: 2%;
-  margin-bottom: 1%;
-  color: #3c4257;
-  height: 4%;
-}
-
-.hidden-button {
-  display: none;
-}
-
-.confirmation-message {
-  width: 90%;
-  text-align: center;
-}
-
-.transparent {
-  opacity: 0;
-}
-
-.dashboard-top-page-number {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  width: 71%;
-  -webkit-box-pack: end;
-  -ms-flex-pack: end;
-  justify-content: flex-end;
-}
-
-#items-per-page {
-  margin-left: 2%;
-}
-
-.items-pp-select {
-  margin-left: 2%;
-}
-
-.column-headers {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-  border: solid #e3e8ee 1px;
-  width: 100%;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  margin-bottom: 0px;
-  border-radius: 5px;
-}
-
-.hidden-input {
-  opacity: 0;
-  text-align: left;
-  margin: auto 1%;
-}
-
-.column-headers-flex {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  text-align: left;
-  -webkit-box-pack: start;
-  -ms-flex-pack: start;
-  justify-content: flex-start;
-  width: 100%;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  color: #3c4257;
-  margin: 0 auto;
-}
-
-.column-headers div h4 {
-  font-weight: bold;
-  padding-top: 0.5%;
-}
-
-.description {
-  margin-left: 0px;
-  text-align: left;
-  width: 45%;
-}
-
-.location {
-  width: 18%;
-  margin-left: 2%;
-}
-
-.date {
-  width: 15%;
-  margin-left: 7%;
-}
-
-.incidents {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-}
-/*# sourceMappingURL=Incidents.css.map */
