--- src/components/AdminDashboard/PendingIncident.css
@@ -1,80 +1,0 @@
-.pending-incident {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: justify;
-      -ms-flex-pack: justify;
-          justify-content: space-between;
-  border: solid #e3e8ee 1px;
-  width: 100%;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  margin-bottom: 0px;
-  border-radius: 5px;
-}
-
-.toggled {
-  border-bottom: none;
-  background-color: #f7fafc;
-}
-
-#incident-description {
-  margin-left: 0px;
-  font-weight: 500;
-  width: 45%;
-  text-align: left;
-  padding: 0px;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.incident-info {
-  text-align: left;
-  padding-top: 1%;
-}
-
-.incident-info-text-wrap {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  text-align: left;
-  -webkit-box-pack: start;
-      -ms-flex-pack: start;
-          justify-content: flex-start;
-  width: 100%;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  color: #3c4257;
-  font-weight: 500;
-  margin: 0 auto;
-}
-
-.incident-info:first-child {
-  text-align: left;
-  margin: auto 1%;
-}
-
-.incident-info:nth-child(2) {
-  width: 18%;
-  margin-left: 1%;
-}
-
-.incident-info:nth-child(3) {
-  width: 15%;
-  margin-left: 7%;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.incident-info:last-child {
-  margin-left: 5%;
-}
-
-.more-info {
-  cursor: pointer;
-}
-/*# sourceMappingURL=PendingIncident.css.map */
